Created separate folder for SQL recording files Updated imports accordingly
