+++ http_server/flask_blueprints/text_sensor_readings.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request
 from operations_modules import logger
 from operations_modules import app_config_access
-from operations_modules.recording_interval import get_interval_sensor_readings
+from sensor_recording_modules.recording_interval import get_interval_sensor_readings
 from sensor_modules import sensor_access
 
 html_sensor_readings_routes = Blueprint("html_sensor_readings_routes", __name__)
