+++ sensor_recording_modules/variance_checks.py
